feat(library): Improve data freshness, display, and structure
This commit addresses several issues related to the media library to improve data freshness, correct display logic, and refactor the underlying code for better maintainability.

- **Fix Stale Data in Library Chooser:**
  The library chooser dialog now fetches the latest images from Firestore every time it is opened. This resolves a critical bug where newly generated images would not appear in the chooser until the app was fully reloaded. The data fetching logic was moved into the `library_chooser_button` component's `open_dialog` handler.

- **Add Manual Refresh to Main Library:**
  A "Refresh" icon button has been added to the main library page. This gives users an explicit and predictable way to reload the media list from Firestore on demand.

- **Fix Display of Batch Images:**
  The library chooser now correctly displays every image from a batch generation, not just the first one. The rendering logic was updated to iterate through the `gcs_uris` list for each media item.

- **Refactor Data Access Logic:**
  The `get_media_for_page` function was moved from `pages/library.py` to the more appropriate `common/metadata.py`. This centralizes the application's data access layer, separating it from the UI and improving the overall architecture.

- **UI Polish:**
  - Fixed an issue where two refresh buttons were erroneously displayed.
  - Corrected the vertical alignment of the icon and title in the page header.
  - Fixed the text alignment of multi-line items in the side navigation.
