+++ experiments/veo-app/config/imagen_models.py
@@ -31,28 +31,28 @@ IMAGEN_MODELS: List[ImagenModelConfig] = [
         display_name="Imagen 3 Fast",
         supported_aspect_ratios=["1:1", "3:4", "4:3", "9:16", "16:9"],
         max_samples=4,
-        default_samples=1,
+        default_samples=4,
     ),
     ImagenModelConfig(
         model_name="imagen-3.0-generate-002",
         display_name="Imagen 3",
         supported_aspect_ratios=["1:1", "3:4", "4:3", "9:16", "16:9"],
         max_samples=4,
-        default_samples=1,
+        default_samples=4,
     ),
     ImagenModelConfig(
         model_name="imagen-4.0-generate-preview-06-06",
         display_name="Imagen 4 (preview)",
         supported_aspect_ratios=["1:1", "3:4", "4:3", "9:16", "16:9"],
         max_samples=4,
-        default_samples=1,
+        default_samples=4,
     ),
     ImagenModelConfig(
         model_name="imagen-4.0-fast-generate-preview-06-06",
         display_name="Imagen 4 Fast (preview)",
         supported_aspect_ratios=["1:1", "3:4", "4:3", "9:16", "16:9"],
         max_samples=4,
-        default_samples=1,
+        default_samples=4,
     ),
     ImagenModelConfig(
         model_name="imagen-4.0-ultra-generate-preview-06-06",
